e_module: Add event handler for ECORE_WL_EVENT_INTERFACES_BOUND
Summary:
- Add event handler to make sure e_module_all_load() done after wayland compositor interfaces are bound
- Delete the "return" code if still loading modules, since "wl_desktop_shell" module create wl-shell and xdg-shell interfaces needed to be bound.

Test Plan:
(1) Run enlightenment on wayland
(2) If e_module attemps to make the "Error loading Module" popup window, you will see deadlock waiting for the wayland sync "done" event to be received. This means the wayland global interfaces including wl.compositor is not ready.

below is the mainloop we are blocked
---------------------------------------------------------------
_ecore_wl_init_wait(void)
{
   while (!_ecore_wl_disp->init_done)
     wl_display_dispatch(_ecore_wl_disp->wl.display);
}

Change-Id: Id1207dfdf40b0fe078ed9700c63d1203fb26ec5e
Signed-off-by: Duna Oh <duna.oh@samsung.com>