e_pixmap: manage deleted pixmap hash table for some pixmaps that will be freed soon
when enlightenment try to create new e_pixmap, e_pixmap hash table is checked
first with surface resource id as a hash key. if there is a pixmap data that
was registered by given key, e_pixmap is NOT newly created but its reference
count is just increased.

This mechanism can generate problem when more than two successive surface
resources are allocated same address. enlightenment have delayed free of e_pixmap
because there are a lot of thing to do before the pixmap is freed. so interval
between removal of a e_pixmap for previous surface resource and creation of
e_pixmap for next surface resource can be made. When the interval is
made and latest surface resource is allocated same address with previous
surface resource, latest surface resource fails to create new e_pixmap
and it will use e_pixmap for previous surface and it can cause abuse of
memory.

To fix this problem, hash table for deleted pixmap is newly added.

Change-Id: I54ead4101210e3dda75d03d18a1e3e840dd8813b
Signed-off-by: MinJeong Kim <minjjj.kim@samsung.com>