Fixed to render image with previous size
Set size of image object to new buffer size when new buffer resource be
set. This patch fixes a problem to draw image object with size of previous
buffer that not rendered.

The problem mentioned above is derived from current buffer management process.
because the process doesn't ensure to render all committed buffer, it means
there are buffers never rendered(even if those buffers changes size of image
object already it's possible to be skip rendering). In the future, new process
is considerable that manage new attached buffer resources with pending list
if previous buffer is not rendered yet to resolve the skip problem.

Change-Id: Ia4bc7296c9770a1a2a28f7872f1c7ef1225a712b
Signed-off-by: MinJeong Kim <minjjj.kim@samsung.com>