e_comp_hwc: native surface set at the deactivated.
At the time of the transition from the nocomposite to the composite mode,
the last buffer commited at the nocompiste has to set the native surface
to the evas object of the cw. The hwc makes the copied surface and sets
the native surface set with the tbm surface type to the evas object.
The copied surface is destoryed when the next buffer commit of the wl_surface
is done by the client.

Change-Id: I44d40f4c6b7304871a1e058302d7410f4ccdc7cd