e_pixmap: ref/unref wl_shm_pool for safe accessing shm buffer data
To manage unmapping of shm buffer pool only inside of wayland-shm
implementation is totally unsafe because there are some cases that the wayland
display server handles data of some shm buffer after destroying of a wayland
client which wass wrtier of the shm buffer. So ref/unref APIs of wl_shm_pool
are necessary to use buffer without risk.

Change-Id: I7ae9251a9755ebad97afb8d6fd8e2d7f05fa9a22