fix the subsurface's wrong behavior
A sub-surface becomes mapped, when a non-NULL wl_buffer is applied and
the parent surface is mapped. The order of which one happens first is
irrelevant. A sub-surface is hidden if the parent becomes hidden, or if
a NULL wl_buffer is applied. These rules apply recursively through the
tree of surfaces.

If the parent wl_surface object is destroyed, the sub-surface is unmapped.

Change-Id: I255593b01fe5e456ae5e0758a305054c9e79dc72